# docs/mobile/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/student/dashboard": {
      "get": {
        "operationId": "StudentController_getDashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get student dashboard data",
        "tags": [
          "Student"
        ]
      }
    },
    "/student/courses": {
      "get": {
        "operationId": "StudentController_getMyCourses",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my enrolled courses",
        "tags": [
          "Student"
        ]
      }
    },
    "/student/courses/{courseId}/check-enrollment": {
      "get": {
        "operationId": "StudentController_checkEnrollment",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check if student is enrolled in a course",
        "tags": [
          "Student"
        ]
      }
    },
    "/student/explore": {
      "get": {
        "operationId": "StudentController_getAvailableCourses",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get available (grade-filtered) courses",
        "tags": [
          "Student"
        ]
      }
    },
    "/student/courses/{slug}/workspace": {
      "get": {
        "operationId": "StudentController_getCourseWorkspace",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get course learning workspace content",
        "tags": [
          "Student"
        ]
      }
    },
    "/student/lessons/{lessonId}/note": {
      "get": {
        "operationId": "StudentController_getLessonNote",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get student note for a lesson",
        "tags": [
          "Student"
        ]
      },
      "post": {
        "operationId": "StudentController_saveLessonNote",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Save student note for a lesson",
        "tags": [
          "Student"
        ]
      }
    },
    "/student/lessons/{lessonId}/complete": {
      "post": {
        "operationId": "StudentController_markLessonCompleted",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a lesson as completed",
        "tags": [
          "Student"
        ]
      }
    },
    "/student/notifications": {
      "get": {
        "operationId": "StudentController_getNotifications",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my notifications",
        "tags": [
          "Student"
        ]
      }
    },
    "/student/notifications/{id}/read": {
      "post": {
        "operationId": "StudentController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark notification as read",
        "tags": [
          "Student"
        ]
      }
    },
    "/student/video/{videoId}/stream": {
      "get": {
        "operationId": "StudentController_getVideoStream",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get video stream URL (requires enrollment)",
        "tags": [
          "Student"
        ]
      }
    },
    "/student/video/{videoId}/duration": {
      "patch": {
        "operationId": "StudentController_updateVideoDuration",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Silently update video duration if it is currently 0",
        "tags": [
          "Student"
        ]
      }
    },
...
  },
  "info": {
    "title": "Masarak API",
    "description": "The Masarak Enterprise E-Learning API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}